+++ backend/src/main/java/edonymyeon/backend/post/application/PostReadService.java
@@ -8,7 +8,12 @@ import edonymyeon.backend.image.domain.Domain;
 import edonymyeon.backend.member.application.dto.MemberId;
 import edonymyeon.backend.member.domain.Member;
 import edonymyeon.backend.member.repository.MemberRepository;
-import edonymyeon.backend.post.application.dto.*;
+import edonymyeon.backend.post.application.dto.response.AllThumbsInPostResponse;
+import edonymyeon.backend.post.application.dto.response.GeneralPostInfoResponse;
+import edonymyeon.backend.post.application.dto.response.ReactionCountResponse;
+import edonymyeon.backend.post.application.dto.response.SpecificPostInfoResponse;
+import edonymyeon.backend.post.application.dto.response.ThumbsStatusInPostResponse;
+import edonymyeon.backend.post.application.dto.response.WriterDetailResponse;
 import edonymyeon.backend.post.domain.HotPostPolicy;
 import edonymyeon.backend.post.domain.Post;
 import edonymyeon.backend.post.repository.PostRepository;
