+++ backend/src/main/java/edonymyeon/backend/post/application/PostThumbsService.java
@@ -1,8 +1,8 @@
 package edonymyeon.backend.post.application;
 
 import edonymyeon.backend.member.application.dto.MemberId;
-import edonymyeon.backend.post.application.dto.AllThumbsInPostResponse;
-import edonymyeon.backend.post.application.dto.ThumbsStatusInPostResponse;
+import edonymyeon.backend.post.application.dto.response.AllThumbsInPostResponse;
+import edonymyeon.backend.post.application.dto.response.ThumbsStatusInPostResponse;
 
 public interface PostThumbsService {
