--- backend/src/main/java/edonymyeon/backend/post/application/dto/PostResponse.java
@@ -1,5 +1,0 @@
-package edonymyeon.backend.post.application.dto;
-
-public record PostResponse(Long id) {
-
-}
