--- backend/src/main/java/edonymyeon/backend/post/application/dto/WriterResponse.java
@@ -1,5 +1,0 @@
-package edonymyeon.backend.post.application.dto;
-
-public record WriterResponse(String nickName) {
-
-}
