+++ backend/src/main/java/edonymyeon/backend/post/application/dto/request/PostModificationRequest.java
@@ -1,14 +1,14 @@
-package edonymyeon.backend.post.application.dto;
+package edonymyeon.backend.post.application.dto.request;
 
 import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
 
-public record PostRequest(
+public record PostModificationRequest(
         String title,
         String content,
         Long price,
+        List<String> originalImages,
         List<MultipartFile> newImages
 ) {
 
 }
-
