+++ backend/src/main/java/edonymyeon/backend/post/application/dto/request/PostRequest.java
@@ -1,14 +1,14 @@
-package edonymyeon.backend.post.application.dto;
+package edonymyeon.backend.post.application.dto.request;
 
 import java.util.List;
 import org.springframework.web.multipart.MultipartFile;
 
-public record PostModificationRequest(
+public record PostRequest(
         String title,
         String content,
         Long price,
-        List<String> originalImages,
         List<MultipartFile> newImages
 ) {
 
 }
+
