+++ backend/src/main/java/edonymyeon/backend/post/application/dto/response/AllThumbsInPostResponse.java
@@ -1,4 +1,4 @@
-package edonymyeon.backend.post.application.dto;
+package edonymyeon.backend.post.application.dto.response;
 
 public record AllThumbsInPostResponse(int thumbsUpCount, int thumbsDownCount) {
