+++ backend/src/main/java/edonymyeon/backend/post/application/dto/response/PostResponse.java
@@ -1,0 +1,5 @@
+package edonymyeon.backend.post.application.dto.response;
+
+public record PostResponse(Long id) {
+
+}
