+++ backend/src/main/java/edonymyeon/backend/post/application/dto/response/ReactionCountResponse.java
@@ -1,4 +1,4 @@
-package edonymyeon.backend.post.application.dto;
+package edonymyeon.backend.post.application.dto.response;
 
 public record ReactionCountResponse(int viewCount, int commentCount, int scrapCount) {
