+++ backend/src/main/java/edonymyeon/backend/post/application/dto/response/ThumbsStatusInPostResponse.java
@@ -1,4 +1,4 @@
-package edonymyeon.backend.post.application.dto;
+package edonymyeon.backend.post.application.dto.response;
 
 public record ThumbsStatusInPostResponse(boolean isUp, boolean isDown) {
