+++ backend/src/main/java/edonymyeon/backend/post/application/dto/response/WriterDetailResponse.java
@@ -1,4 +1,4 @@
-package edonymyeon.backend.post.application.dto;
+package edonymyeon.backend.post.application.dto.response;
 
 public record WriterDetailResponse(long id, String nickname, String profileImage) {
