+++ backend/src/main/java/edonymyeon/backend/post/ui/PostReadController.java
@@ -6,8 +6,8 @@ import edonymyeon.backend.post.application.GeneralFindingCondition;
 import edonymyeon.backend.post.application.HotFindingCondition;
 import edonymyeon.backend.post.application.PostReadService;
 import edonymyeon.backend.post.application.PostSlice;
-import edonymyeon.backend.post.application.dto.GeneralPostInfoResponse;
-import edonymyeon.backend.post.application.dto.SpecificPostInfoResponse;
+import edonymyeon.backend.post.application.dto.response.GeneralPostInfoResponse;
+import edonymyeon.backend.post.application.dto.response.SpecificPostInfoResponse;
 import edonymyeon.backend.post.ui.annotation.HotPostSizing;
 import edonymyeon.backend.post.ui.annotation.PostPaging;
 import lombok.RequiredArgsConstructor;
