+++ backend/src/main/java/edonymyeon/backend/thumbs/application/PostThumbsServiceImpl.java
@@ -3,8 +3,8 @@ package edonymyeon.backend.thumbs.application;
 import edonymyeon.backend.member.application.dto.AnonymousMemberId;
 import edonymyeon.backend.member.application.dto.MemberId;
 import edonymyeon.backend.post.application.PostThumbsService;
-import edonymyeon.backend.post.application.dto.AllThumbsInPostResponse;
-import edonymyeon.backend.post.application.dto.ThumbsStatusInPostResponse;
+import edonymyeon.backend.post.application.dto.response.AllThumbsInPostResponse;
+import edonymyeon.backend.post.application.dto.response.ThumbsStatusInPostResponse;
 import edonymyeon.backend.thumbs.domain.AllThumbsInPost;
 import edonymyeon.backend.thumbs.domain.Thumbs;
 import edonymyeon.backend.thumbs.repository.ThumbsRepository;
